Implement basic search agent using BFS (BUG)
BUG : Only works for --layout small.

The BFS implementation of the search agent uses
queues and sets rather than doing things recursively, mainly
to contain most of the code inside one function call, namely
get_action().

Checking legal actions individually  and each successor emerging
from that action using getPacmanSuccesor(action) is better
than creating all successors at once using getPacmanSuccessors
because it allows to add each one to the closed_set variable
individually.

Note that what is added to the closed_set is the successor state
and not the position. A position and a state differ, among other things,
by the directions of the agent which is not included inside the position
variable.
Although using the position in the small layout is optimal, this
only works because there is only one food source and Pacman never has
to go back to the same position twice. This is not the case for the
bigger layouts.