Revise the BFS search agent
The tree generated is now only from the original state to the first
encountered food item. After the tree has been build,
we compute a path from the original state to the goal state and
list all the actions needed to reach that state.
Once that state is reached, we repeat the process.

Instead of getting the legal actions and looking for the successor
states using getSuccessor(), the getPacmanSuccessors() method
is used which is NECESSARY because it allows the total amount of
expanded nodes to be calculated.

Furthermore, the visited set now contains the position and the number of
total food items in the maze, and not the state.
This is due to the fact that pacman can only go back to a certain
position if he ate one food item in the meantime.