Optimize BFS algorithm
Instead of using getNumFood() to mark the current node as visited,
the whole array returned from the getFood() function used. That is
better because one food can be eaten from multiple positions and that
is not reflected in the getNumFood() return value.
Moreover, the food array as well as the position is hashed before
inserting into the visited set, to save on space.