Implement the heuristic "manhattan_sum"
Try to use the heuristic manhattan_minimum but it takes too long
with the medium grid.
Then, implement manhattan_maximum and correct the inital value
of max_man to 0 (not +inf). Works well for all grids.
Finally, implement manhattan_sum that gets the best results but
not sure that this way of doing is admissible.