Add get_info() method, add info to visited in max_value() and min_value()
Remove depth notion as this belongs to hminimax

However, in min-value() do not check that the ghost go back to the same position
as some ghost do by construction. Instead, just add the node to the visited set.
In minimax_decision() and max_value(), add the info to the visited set and check
if not already visited.